Predictive performance of EuroSCORE II

This data set contains estimates on the predictive performance of the European system for cardiac operative risk evaluation (EuroSCORE II) in patients undergoing cardiac surgery. Results are based on the original development study and 22 validations identified by Guida et al.

Usage

data("EuroSCORE")

Format

A data frame with 23 observations on the following 13 variables.

Study

a vector with the first author of each validation study

n

a numeric vector with the total number of patients on which performance estimates are based

n.events

a numeric vector with the total number of observed events

c.index

a numeric vector with the estimated concordance statistic of each validation

se.c.index

a numeric vector with the standard error of the concordance statistics

c.index.95CIl

a numeric vector with the lower bound of the 95% confidence interval of the estimated concordance statistics

c.index.95CIu

a numeric vector with the upper bound of the 95% confidence interval of the estimated concordance statistics

Po

a numeric vector with the overall observed event probability of each validation

Pe

a numeric vector with the overall expected event probability of each validation

SD.Pe

a numeric vector with the standard error of Pe

e.events

a numeric vector with the total number of expected events in each validation

multicentre

a logical vector describing whether the study was a multicentre study

mean.age

a numeric vector describing the mean age of the patients

sd.age

a numeric vector with the spread of the age of the patients

pts.before.2010

a logical vector describing whether studies included patients before 2010 (i.e., before EuroSCORE II was developed)

Details

Published in 2012, EuroSCORE II was developed using logistic regression in a dataset comprising 16,828 adult patients undergoing major cardiac surgery from 154 hospitals in 43 countries over a 12-week period (May-July) in 2010. EuroSCORE II was developed to predict in-hospital mortality for patients undergoing any type of cardiac surgery. In 2014, a systematic review of published evidence on the performance value of the euroSCORE II was undertaken by Guida et al. Twenty-two validations, including more 145,592 patients from 21 external validation articles (one study included two validations) and a split-sample validation contained within original development article were included in the review; 23 validation studies in total.

Source

Guida P, Mastro F, Scrascia G, Whitlock R, Paparella D. Performance of the European System for Cardiac Operative Risk Evaluation II: a meta-analysis of 22 studies involving 145,592 cardiac surgery procedures. J Thorac Cardiovasc Surg. 2014; 148(6):3049–3057.e1.

Nashef SAM, Roques F, Sharples LD, Nilsson J, Smith C, Goldstone AR, et al. EuroSCORE II. Eur J Cardiothorac Surg. 2012; 41(4):734-744; discussion 744-745.

Examples

data(EuroSCORE)