Plot the prior and posterior distribution of a meta-analysis model

Function to generate plots for the prior and posterior distribution of a Bayesian meta-analysis.

Usage

# S3 method for class 'uvmeta'
dplot(x, par, distr_type, plot_type = "dens", ...)

Arguments

x

An object of class "uvmeta"

par

Character string to specify for which parameter a plot should be generated. Options are "mu" (mean of the random effects model) and "tau" (standard deviation of the random effects model).

distr_type

Character string to specify whether the prior distribution ("prior") or posterior distribution ("posterior") should be displayed.

plot_type

Character string to specify whether a density plot ("dens") or histogram ("hist") should be displayed.

...

Additional arguments which are currently not used

Value

An object of class ggplot

Author

Thomas Debray <thomas.debray@gmail.com>

Examples

if (FALSE) { # \dontrun{
data(Roberts)

fit <- with(Roberts, uvmeta(r=SDM, r.se=SE, method="BAYES"))

dplot(fit)
dplot(fit, distr_type = "posterior")
dplot(fit, par = "tau", distr_type = "prior")
dplot(fit, plot_type = "hist")
} # }