Extract Model Fitted Values

Extract the fitted values of a metapred object. By default returns fitted values of the model in the cross-validation procedure, i.e., the predicted values for the validation folds.

Usage

# S3 method for class 'metapred'
fitted(
  object,
  select = "cv",
  step = NULL,
  model = NULL,
  as.stratified = TRUE,
  type = "response",
  ...
)

Arguments

object

object of class metapred

select

character. Select fitted values from "cv" (default) or from "global" model.

step

character or numeric. Name or number of step to select if select = "cv". Defaults to best step.

model

character or numeric. Name or number of model to select if select = "cv". Defaults to best model.

as.stratified

logical. Determines whether returned predictions are stratified in a list (TRUE, default) or in their original order (FALSE).

type

character. Type of fitted value.

...

For compatibility only.

Details

Function still under development, use with caution.

Only returns type = "response".

Author

Valentijn de Jong