Gelman-Rubin-Brooks plot

This plot shows the evolution of Gelman and Rubin's shrink factor as the number of iterations increases. The code is adapted from the R package coda.

Usage

# S3 method for class 'valmeta'
gelmanplot(x, confidence = 0.95, ...)

Arguments

x

An mcmc object

confidence

The coverage probability of the confidence interval for the potential scale reduction factor

...

Additional arguments which are currently not used

Value

A ggplot object.

An object of class ggplot

Author

Thomas Debray <thomas.debray@gmail.com>

Examples

if (FALSE) { # \dontrun{
data(EuroSCORE)

# Meta-analysis of the concordance statistic
fit <- valmeta(cstat=c.index, cstat.se=se.c.index, cstat.cilb=c.index.95CIl,
               cstat.ciub=c.index.95CIu, N=n, O=n.events, 
               data=EuroSCORE, method="BAYES", slab=Study)
gelmanplot(fit)
} # }