Impute missing values by their conditional mean

This function imputes missing values by their conditional mean

Usage

impute_conditional_mean(x, mu, Sigma)

Arguments

x

A vector with observations, some of which may be missing (indicated by NA)

mu

A vector with the population means for 'x'. No missing values are allowed here.

Sigma

A matrix describing the population covariance of 'x'

Value

A vector where missing values for 'x' have been replaced by their conditional mean

Author

Thomas Debray <thomas.debray@gmail.com>

Examples

# Define the population means
mu <- c(0, 1, 2)

# Define the covariance of the population
Sigma <- diag(1,3)
Sigma[1,2] <- Sigma[2,1] <- 0.3 
Sigma[2,3] <- Sigma[3,2] <- 0.1
Sigma[1,3] <- Sigma[3,1] <- -0.2

# Generate a 'random' sample from the population that is partially observed
x <- c(NA, 2, 4)

# Impute the missing values
impute_conditional_mean (x=x, mu=mu, Sigma=Sigma)
#> [1] -0.1414141  2.0000000  4.0000000