Print the log-likelihood

This function provides the (restricted) log-likelihood of a fitted model.

Usage

# S3 method for riley
logLik(object, ...)

Arguments

object

A riley object, representing a fitted alternative model for bivariate random-effects meta-analysis when the within-study correlations are unknown.

...

Additional arguments to be passed on to other functions, currently ignored.

Value

Returns an object of class logLik. This is the (restricted) log-likelihood of the model represented by object evaluated at the estimated coefficients. It contains at least one attribute, "df" (degrees of freedom), giving the number of (estimated) parameters in the model.

References

Riley RD, Thompson JR, Abrams KR. An alternative model for bivariate random-effects meta-analysis when the within-study correlations are unknown. Biostatistics 2008; 9: 172--186.

Author

Thomas Debray <thomas.debray@gmail.com>

Examples

data(Daniels)
fit <- riley(Daniels,control=list(maxit=10000))
logLik(fit)
#> 'log Lik.' -48.85119 (df=5)