Class "uvmeta". Result of a univariate meta-analysis.

This class encapsulates results of a univariate meta-analysis.

Objects from the Class

Objects can be created by calls of the form uvmeta.

Slots

call:

(language) The call to uvmeta.

data:

(data frame) The data used for the meta-analysis.

results:

(data frame) Contains the pooled effect size (mu), the between-study variability (tausq), Cochran's Q statistic (Q) and Higgins' and Thompson's I square statistic (Isq). For each estimate, error variances are provided with predefined confidence (method="MOM") or credibility (method="bayes") intervals.

model:

(character) The meta-analysis model used.

method:

(character) The estimator used.

na.action:

(character) Information from the action which was applied to object if NAs were handled specially, or NULL.

df:

(numeric) Degrees of freedom.

numstudies:

(numeric) The amount of studies used in the meta-analysis.

pred.int:

(data frame) A prediction interval, predicting in what range future effect sizes will fall given what has already been observed (based on a Student's t-distribution, cfr. Riley 2011)

formula:

(character) If a formula was specified, a character vector giving the formula and parameter specifications.

Methods

print

signature(object = "uvmeta"): Print object summary.

forest

signature(object = "uvmeta"): Plot a forest plot with the summary estimate.

summary

signature(object = "uvmeta"): Generate object summary.

Examples

data(Collins)

#Extract effect size and error variance
r <- Collins$logOR
vars <- Collins$SE**2

#Frequentist random-effects meta-analysis
fit1 <- uvmeta(r,vars)

#Extract results
fit1$results
#> NULL