Hypothetical dataset for diagnosis of Deep Vein Thrombosis (DVT)

A hypothetical dataset with 500 subjects suspected of having deep vein thrombosis (DVT).

Usage

data(DVTipd)

Format

A data frame with 500 observations of 16 variables.

sex

gender (0=female, 1=male)

malign

active malignancy (0=no active malignancy, 1=active malignancy)

par

paresis (0=no paresis, 1=paresis)

surg

recent surgery or bedridden

tend

tenderness venous system

oachst

oral contraceptives or hst

leg

entire leg swollen

notraum

absence of leg trauma

calfdif3

calf difference >= 3 cm

pit

pitting edema

vein

vein distension

altdiagn

alternative diagnosis present

histdvt

history of previous DVT

ddimdich

dichotimized D-dimer value

dvt

final diagnosis of DVT

study

study indicator

Details

Hypothetical dataset derived from the Individual Participant Data Meta-Analysis from Geersing et al (2014). The dataset consists of consecutive outpatients with suspected deep vein thrombosis, with documented information on the presence or absence of proximal deep vein thrombosis (dvt) by an acceptable reference test. Acceptable such tests were either compression ultrasonography or venography at initial presentation, or, if venous imaging was not performed, an uneventful follow-up for at least three months.

Source

Geersing GJ, Zuithoff NPA, Kearon C, Anderson DR, Ten Cate-Hoek AJ, Elf JL, et al. Exclusion of deep vein thrombosis using the Wells rule in clinically important subgroups: individual patient data meta-analysis. BMJ. 2014;348:g1340.

Examples

data(DVTipd)
str(DVTipd) 
#> 'data.frame':	500 obs. of  16 variables:
#>  $ sex     : num  0 1 0 1 0 0 1 0 1 0 ...
#>  $ malign  : num  0 0 0 0 0 0 0 0 0 0 ...
#>  $ par     : num  0 0 1 0 0 0 0 0 0 0 ...
#>  $ surg    : num  0 0 0 0 0 0 0 0 1 0 ...
#>  $ tend    : num  1 1 0 1 1 0 0 1 1 1 ...
#>  $ oachst  : num  0 0 0 0 0 0 0 0 0 0 ...
#>  $ leg     : num  1 0 0 0 0 1 1 0 0 0 ...
#>  $ notraum : num  1 1 1 1 1 0 0 1 0 1 ...
#>  $ calfdif3: num  0 0 0 0 0 0 0 0 0 0 ...
#>  $ pit     : num  0 0 0 0 0 1 0 1 1 1 ...
#>  $ vein    : num  0 0 0 0 1 0 0 0 0 1 ...
#>  $ altdiagn: num  1 0 1 1 1 0 1 1 1 1 ...
#>  $ histdvt : num  0 1 0 0 0 0 1 0 0 0 ...
#>  $ ddimdich: num  1 0 0 0 0 1 1 0 1 1 ...
#>  $ dvt     : num  0 0 0 0 0 0 0 0 0 0 ...
#>  $ study   : Factor w/ 4 levels "a","b","c","d": 1 4 1 4 1 4 4 4 4 2 ...
summary(apply(DVTipd,2,as.factor))
#>      sex               malign              par                surg          
#>  Length:500         Length:500         Length:500         Length:500        
#>  Class :character   Class :character   Class :character   Class :character  
#>  Mode  :character   Mode  :character   Mode  :character   Mode  :character  
#>      tend              oachst              leg              notraum         
#>  Length:500         Length:500         Length:500         Length:500        
#>  Class :character   Class :character   Class :character   Class :character  
#>  Mode  :character   Mode  :character   Mode  :character   Mode  :character  
#>    calfdif3             pit                vein             altdiagn        
#>  Length:500         Length:500         Length:500         Length:500        
#>  Class :character   Class :character   Class :character   Class :character  
#>  Mode  :character   Mode  :character   Mode  :character   Mode  :character  
#>    histdvt            ddimdich             dvt               study          
#>  Length:500         Length:500         Length:500         Length:500        
#>  Class :character   Class :character   Class :character   Class :character  
#>  Mode  :character   Mode  :character   Mode  :character   Mode  :character  

## Develop a prediction model to predict presence of DVT
model.dvt <- glm("dvt~sex+oachst+malign+surg+notraum+vein+calfdif3+ddimdich", 
                  family=binomial, data=DVTipd)
summary(model.dvt)
#> 
#> Call:
#> glm(formula = "dvt~sex+oachst+malign+surg+notraum+vein+calfdif3+ddimdich", 
#>     family = binomial, data = DVTipd)
#> 
#> Coefficients:
#>             Estimate Std. Error z value Pr(>|z|)    
#> (Intercept)  -5.1664     0.6365  -8.117 4.76e-16 ***
#> sex           0.8146     0.2825   2.883  0.00393 ** 
#> oachst        0.4324     0.6227   0.694  0.48739    
#> malign        0.5679     0.4025   1.411  0.15826    
#> surg          0.1002     0.4111   0.244  0.80734    
#> notraum       0.3351     0.3700   0.906  0.36513    
#> vein          0.4831     0.3186   1.516  0.12939    
#> calfdif3      1.1841     0.2819   4.200 2.67e-05 ***
#> ddimdich      2.6081     0.5310   4.911 9.04e-07 ***
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 
#> (Dispersion parameter for binomial family taken to be 1)
#> 
#>     Null deviance: 446.24  on 499  degrees of freedom
#> Residual deviance: 345.98  on 491  degrees of freedom
#> AIC: 363.98
#> 
#> Number of Fisher Scoring iterations: 6
#>