Predictive performance of the Framingham Risk Score in male populations

This data set contains estimates on the performance of the Framingham model for predicting coronary heart disease in male populations (Wilson 1998). Results are based on the original development study and 20 validations identified by Damen et al (BMC Med, 2017).

Usage

data("Framingham")

Format

A data frame with 24 observations on the following 19 variables.

AuthorYear

a vector describing the study authors

n

a numeric vector with the total number of patients on which performance estimates are based

n.events

a numeric vector with the total number of observed events

c.index

a numeric vector with the estimated concordance statistic of each validation

se.c.index

a numeric vector with the standard error of the concordance statistics

c.index.95CIl

a numeric vector with the lower bound of the 95% confidence interval of the estimated concordance statistics

c.index.95CIu

a numeric vector with the upper bound of the 95% confidence interval of the estimated concordance statistics

Po

a numeric vector with the overall observed event probability of each validation

Pe

a numeric vector with the overall expected event probability of each validation

t.val

a numeric vector describing the time period in which predictive performance was assessed for each validation

mean_age

a numeric vector describing the mean age of the patients

sd_age

a numeric vector with the spread of the age of the patients

mean_SBP

a numeric vector with the mean systolic blood pressure in the validation studies (mm Hg)

sd_SBP

a numeric vector with the spread of systolic blood pressure in the validation studies

mean_total_cholesterol

a numeric vector with the mean total cholesterol in the validation studies (mg/dL)

sd_total_cholesterol

a numeric vector with the spread of total cholesterol in the validation studies

mean_hdl_cholesterol

a numeric vector with the mean high-density lipoprotein cholesterol in the validation studies (mg/dL)

sd_hdl_cholesterol

a numeric vector with the spread of high-density lipoprotein cholesterol in the validation studies

pct_smoker

a numeric vector with the percentage smokers in the validation studies

Details

The Framingham Risk Score allows physicians to predict 10-year coronary heart disease (CHD) risk in patients without overt CHD. It was developed in 1998 from a middle-aged white population sample, and has subsequently been validated across different populations. The current dataset contains the original (internal validation) results, as well as 23 external validations which were identified through a systematic review. In this review, studies were eligible for inclusion if they described the validation of the original Framingham model and assessed its performance for fatal or nonfatal CHD in males from a general population setting.

Source

Damen JAAG, Hooft L, Schuit E, Debray TPA, Collins GS, Tzoulaki I, et al. Prediction models for cardiovascular disease risk in the general population: systematic review. BMJ. 2016;i2416.

Damen JAAG, Pajouheshnia R, Heus P, Moons KGM, Reitsma JB, Scholten RJPM, et al. Performance of the Framingham risk models and Pooled Cohort Equations: a systematic review and meta-analysis. BMC Med. 2017;17(1):109.

Wilson PW, D'Agostino RB, Levy D, Belanger AM, Silbershatz H, Kannel WB. Prediction of coronary heart disease using risk factor categories. Circulation. 1998; 97(18):1837–47.

Examples

data(Framingham)