Forest plot

Generate a forest plot by specifying the various effect sizes, confidence intervals and summary estimate.

Usage

# Default S3 method
forest(
  theta,
  theta.ci.lb,
  theta.ci.ub,
  theta.slab,
  theta.summary,
  theta.summary.ci.lb,
  theta.summary.ci.ub,
  theta.summary.pi.lb,
  theta.summary.pi.ub,
  title,
  sort = "asc",
  theme = theme_bw(),
  predint.linetype = 1,
  xlim,
  xlab = "",
  refline = 0,
  label.summary = "Summary Estimate",
  label.predint = "Prediction Interval",
  nrows.before.summary = 1,
  study.digits = 2,
  study.shape = 15,
  col.diamond = "white",
  col.predint = "black",
  size.study = 0.5,
  size.predint = 1,
  lty.ref = "dotted",
  ...
)

Arguments

theta

Numeric vector with effect size for each study

theta.ci.lb

Numeric vector specifying the lower bound of the confidence interval of the effect sizes

theta.ci.ub

Numeric vector specifying the upper bound of the confidence interval of the effect sizes

theta.slab

Character vector specifying the study labels

theta.summary

Meta-analysis summary estimate of the effect sizes

theta.summary.ci.lb

Lower bound of the confidence (or credibility) interval of the summary estimate

theta.summary.ci.ub

Upper bound of the confidence (or credibility) interval of the summary estimate

theta.summary.pi.lb

Lower bound of the (approximate) prediction interval of the summary estimate.

theta.summary.pi.ub

Upper bound of the (approximate) prediction interval of the summary estimate.

title

Title of the forest plot

sort

By default, studies are sorted by ascending effect size (sort="asc"). Set to "desc" for sorting in reverse order, or any other value to ignore sorting.

theme

Theme to generate the forest plot. By default, the classic dark-on-light ggplot2 theme is used. See ggtheme for more information.

predint.linetype

The linetype of the prediction interval

xlim

The x limits (x1, x2) of the forest plot

xlab

Optional character string specifying the X label

refline

Optional numeric specifying a reference line

label.summary

Optional character string specifying the label for the summary estimate

label.predint

Optional character string specifying the label for the (approximate) prediction interval

nrows.before.summary

How many empty rows should be introduced between the study results and the summary estimates

study.digits

How many significant digits should be used to print the stuy results

study.shape

Plotting symbol to use for the study results. By default, a filled square is used.

col.diamond

The filling color for the diamond representing the summary estimate. E.g. "red", "blue", or hex color code ("#2e8aff")

col.predint

Line color for the prediction interval. E.g. "red", "blue", or hex color code ("#2e8aff")

size.study

Line width for the study results in mm

size.predint

Line width for the prediction interval in mm

lty.ref

Line type for the reference line

...

Additional arguments, which are currently ignored.

Value

An object of class ggplot

Author

Thomas Debray <thomas.debray@gmail.com>